add 'conditions' to the config file
This is a set() that holds the condition flags for this run.

It is pre-seeded with some flags according to what platform you're
running on (by way of inspecting sys.platform)

It is intended that people should update the conditions in their
jhbuildrc files like so:

  conditions.add('condition')

or

  conditions.discard('condition')

But they can also be replaced outright, as long as it remains a set.

Future commits will introduce behaviour that depends on this setting.

https://bugzilla.gnome.org/show_bug.cgi?id=720839