Add --conditions= commandline argument
This can be used like --conditions=+doc,-wayland in order to modify the
condition set from the commandline.

The user's jhbuildrc is capable of inspecting conditions set from the
commandline by way of checking

    if 'flag' in conditions:
        ...

but the arguments on the commandline always override what is written in
jhbuildrc.  This is done by applying the modifiers twice -- once before
parsing the config and then again, after.

https://bugzilla.gnome.org/show_bug.cgi?id=720839