Handle <if> tags in moduleset files
If we encounter an <if> tag, consult the conditions set in the config
in order to decide if we should include its content or not.  If the
condition is met, the child elements are added to the parent of the
<if/> tag as if the condition tag were not there at all.  If the
condition is not met, the entire content is simply dropped.

We do the processing as a transformation on the DOM as a whole,
immediately after parsing the moduleset XML, before doing any additional
processing.  This allows <if> to be used for anything and it means we
don't need to deal with it separately from each place.

Although the tool itself will accept <if> anywhere we use the schemas to
restrict its use to the purposes of conditionalising dependencies
(including suggests) and {autogen,make,makeinstall}args.

https://bugzilla.gnome.org/show_bug.cgi?id=720839