autotools: collect args from elements as well
The autotools modtype has support for specifying autogenargs, makeargs
and makeinstallargs as attributes on the <autotools/> tag.

Add support for specifying additional arguments via the elements
<autogenargs/>, <makeargs/> and <makeinstallargs/>.  The additional
arguments are concatenated to the list.

This is intended to be used with <conditional> to provide a way to
customise the arguments based on the condition set.

https://bugzilla.gnome.org/show_bug.cgi?id=720839